create simple_ping_pong file and move ping_pong into it from ping_pong.rs
