+++ src/main.rs
@@ -1,6 +1,6 @@
 mod ping_pong;
 
 fn main() {
-    ping_pong::run();
+    ping_pong::simple_ping_pong::run();
 }
