+++ src/ping_pong.rs
@@ -1,48 +1,1 @@
-use std::thread;
-use std::sync::mpsc::{channel, RecvError};
-
-#[derive(Debug)]
-struct Message{
-    pub value: String
-}
-
-pub fn run(){
-    let (ping_sender, ping_receiver) = channel::<Message>();
-    let (pong_sender, pong_receiver) = channel::<Message>();
-    let mut handles = vec![];
-
-    // Ping
-    let ping_handle = thread::spawn(move||{
-        let message = Message{value: String::from("PING")};
-        let send_result = ping_sender.send(message);
-        let pong = pong_receiver.recv();
-        match pong {
-            Ok(p) => {
-                println!("simple mission compleated message = {:?}", p);
-            }
-            Err(e) => {
-                println!("ERROR!  {:?}", e);
-            }
-        }
-    });
-    handles.push(ping_handle);
-
-    // Pong 
-    let pong_hundle = thread::spawn(move||{
-        let received_message = ping_receiver.recv();
-        match  received_message {
-            Ok(m) =>{
-                println!("receive.... {:?}", m);
-                let _ = pong_sender.send(Message{value: String::from("PONG")});
-            }
-            Err(e) => {
-                println!("PONG ERROR");
-            }
-        }
-    });
-    handles.push(pong_hundle);
-
-    for h in handles{
-        let _ = h.join();
-    }
-}
+pub mod simple_ping_pong;
